Handle potential overflow from PrefetchCountCalculation

+++ src/NServiceBus.Transport.RabbitMQ/PrefetchCountCalculation.cs
@@ -5,5 +5,5 @@
     /// </summary>
     /// <param name="maximumConcurrency">Maximum concurrency of the message receiver.</param>
     /// <returns>The prefetch count to use for the receiver.</returns>
-    public delegate int PrefetchCountCalculation(int maximumConcurrency);
+    public delegate long PrefetchCountCalculation(int maximumConcurrency);
 }
